# Backend/analysis/src/schemas/BufferAnalysisOpenAPISchema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "缓冲区分析API",
    "description": "SuperMap缓冲区分析服务接口定义",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
    "/api/buffer-analysis": {
      "post": {
        "summary": "执行缓冲区分析",
        "description": "对指定的地理要素执行缓冲区分析操作",
        "operationId": "executeBufferAnalysis",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BufferAnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "缓冲区分析执行成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BufferAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "图层不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/buffer-analysis/{resultId}": {
      "get": {
        "summary": "获取分析结果",
        "description": "根据结果ID获取缓冲区分析结果",
        "operationId": "getAnalysisResult",
        "parameters": [
          {
            "name": "resultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "分析结果ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取结果成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "结果ID无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "结果不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BufferAnalysisRequest": {
        "type": "object",
        "description": "缓冲区分析请求数据结构",
        "properties": {
          "sourceLayerId": {
            "type": "string",
            "description": "源图层ID，与sourceData二选一"
          },
          "sourceData": {
            "$ref": "#/components/schemas/GeoJSONData",
            "description": "源GeoJSON数据，与sourceLayerId二选一"
          },
          "bufferSettings": {
            "$ref": "#/components/schemas/BufferSettings",
            "description": "缓冲区设置参数"
          },
          "featureFilter": {
            "$ref": "#/components/schemas/FeatureFilter",
            "description": "要素过滤条件"
          },
          "options": {
            "$ref": "#/components/schemas/AnalysisOptions",
            "description": "分析选项设置"
          }
        },
        "required": ["bufferSettings"],
        "oneOf": [
          {
            "required": ["sourceLayerId"]
          },
          {
            "required": ["sourceData"]
          }
        ]
      },
      "BufferSettings": {
        "type": "object",
        "description": "缓冲区设置参数",
        "properties": {
          "radius": {
            "type": "number",
            "minimum": 0.001,
            "maximum": 100000,
            "description": "缓冲距离，单位为米"
          },
          "semicircleLineSegment": {
            "type": "integer",
            "minimum": 1,
            "maximum": 64,
            "default": 10,
            "description": "半圆线段数，控制缓冲区圆弧精度"
          },
          "unit": {
            "type": "string",
            "enum": ["meters", "kilometers", "feet", "miles"],
            "default": "meters",
            "description": "距离单位"
          },
          "unionResults": {
            "type": "boolean",
            "default": true,
            "description": "是否合并重叠的缓冲区结果"
          }
        },
        "required": ["radius"]
      },
      "GeoJSONData": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/FeatureCollection"
          },
          {
            "$ref": "#/components/schemas/Feature"
          }
        ]
      },
      "FeatureCollection": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["FeatureCollection"]
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            }
          }
        },
        "required": ["type", "features"]
      },
      "Feature": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["Feature"]
          },
          "geometry": {
            "$ref": "#/components/schemas/Geometry"
          },
          "properties": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": ["type", "geometry"]
      },
      "Geometry": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/Point"
          },
          {
            "$ref": "#/components/schemas/LineString"
          },
          {
            "$ref": "#/components/schemas/Polygon"
          },
          {
            "$ref": "#/components/schemas/MultiPoint"
          },
          {
            "$ref": "#/components/schemas/MultiLineString"
          },
          {
            "$ref": "#/components/schemas/MultiPolygon"
          }
        ]
      },
      "Point": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["Point"]
          },
          "coordinates": {
            "type": "array",
            "minItems": 2,
            "maxItems": 3,
            "items": {
              "type": "number"
            }
          }
        },
        "required": ["type", "coordinates"]
      },
      "LineString": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["LineString"]
          },
          "coordinates": {
            "type": "array",
            "minItems": 2,
            "items": {
              "type": "array",
              "minItems": 2,
              "maxItems": 3,
              "items": {
                "type": "number"
              }
            }
          }
        },
        "required": ["type", "coordinates"]
      },
      "Polygon": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["Polygon"]
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "array",
              "minItems": 4,
              "items": {
                "type": "array",
                "minItems": 2,
                "maxItems": 3,
                "items": {
                  "type": "number"
                }
              }
            }
          }
        },
        "required": ["type", "coordinates"]
      },
      "MultiPoint": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["MultiPoint"]
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "array",
              "minItems": 2,
              "maxItems": 3,
              "items": {
                "type": "number"
              }
            }
          }
        },
        "required": ["type", "coordinates"]
      },
      "MultiLineString": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["MultiLineString"]
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "array",
              "minItems": 2,
              "items": {
                "type": "array",
                "minItems": 2,
                "maxItems": 3,
                "items": {
                  "type": "number"
                }
              }
            }
          }
        },
        "required": ["type", "coordinates"]
      },
      "MultiPolygon": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["MultiPolygon"]
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "array",
                "minItems": 4,
                "items": {
                  "type": "array",
                  "minItems": 2,
                  "maxItems": 3,
                  "items": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "required": ["type", "coordinates"]
      },
      "FeatureFilter": {
        "type": "object",
        "description": "要素过滤条件",
        "properties": {
          "featureIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "指定要素ID列表"
          },
          "spatialFilter": {
            "$ref": "#/components/schemas/SpatialFilter",
            "description": "空间过滤条件"
          }
        }
      },
      "SpatialFilter": {
        "type": "object",
        "description": "空间过滤条件",
        "properties": {
          "bounds": {
            "type": "array",
            "minItems": 4,
            "maxItems": 4,
            "items": {
              "type": "number"
            },
            "description": "边界框 [minLon, minLat, maxLon, maxLat]"
          }
        }
      },
      "AnalysisOptions": {
        "type": "object",
        "description": "分析选项设置",
        "properties": {
          "returnGeometry": {
            "type": "boolean",
            "default": true,
            "description": "是否返回几何信息"
          },
          "returnProperties": {
            "type": "boolean",
            "default": true,
            "description": "是否返回属性信息"
          },
          "resultLayerName": {
            "type": "string",
            "default": "缓冲区分析结果",
            "description": "结果图层名称"
          }
        }
      },
      "BufferAnalysisResponse": {
        "type": "object",
        "description": "缓冲区分析响应结果",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "操作是否成功"
          },
          "data": {
            "$ref": "#/components/schemas/BufferAnalysisResult"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "响应时间戳"
          }
        },
        "required": ["success", "data", "message", "timestamp"]
      },
      "BufferAnalysisResult": {
        "type": "object",
        "description": "缓冲区分析结果数据",
        "properties": {
          "resultId": {
            "type": "string",
            "description": "结果唯一标识"
          },
          "resultName": {
            "type": "string",
            "description": "结果名称"
          },
          "sourceLayerName": {
            "type": "string",
            "description": "源图层名称"
          },
          "bufferSettings": {
            "$ref": "#/components/schemas/BufferSettingsResult"
          },
          "statistics": {
            "$ref": "#/components/schemas/AnalysisStatistics"
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "description": "缓冲区要素结果集"
          },
          "executionTime": {
            "type": "string",
            "description": "执行耗时"
          }
        },
        "required": ["resultId", "resultName", "sourceLayerName", "bufferSettings", "statistics", "features", "executionTime"]
      },
      "BufferSettingsResult": {
        "type": "object",
        "description": "实际使用的缓冲区设置",
        "properties": {
          "radius": {
            "type": "number",
            "description": "缓冲距离"
          },
          "semicircleLineSegment": {
            "type": "integer",
            "description": "半圆线段数"
          },
          "unit": {
            "type": "string",
            "description": "距离单位"
          }
        },
        "required": ["radius", "semicircleLineSegment", "unit"]
      },
      "AnalysisStatistics": {
        "type": "object",
        "description": "分析统计信息",
        "properties": {
          "inputFeatureCount": {
            "type": "integer",
            "description": "输入要素数量"
          },
          "outputFeatureCount": {
            "type": "integer",
            "description": "输出要素数量"
          },
          "totalArea": {
            "type": "number",
            "description": "总面积（平方千米）"
          },
          "averageArea": {
            "type": "number",
            "description": "平均面积（平方千米）"
          }
        },
        "required": ["inputFeatureCount", "outputFeatureCount", "totalArea", "averageArea"]
      },
      "AnalysisResultResponse": {
        "type": "object",
        "description": "分析结果查询响应",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "操作是否成功"
          },
          "data": {
            "type": "object",
            "properties": {
              "resultId": {
                "type": "string",
                "description": "结果ID"
              },
              "message": {
                "type": "string",
                "description": "响应消息"
              },
              "note": {
                "type": "string",
                "description": "备注信息"
              }
            },
            "required": ["resultId", "message"]
          }
        },
        "required": ["success", "data"]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "错误响应格式",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false],
            "description": "操作失败标识"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "错误代码"
              },
              "message": {
                "type": "string",
                "description": "错误消息"
              },
              "details": {
                "type": "string",
                "description": "错误详情（仅开发环境）"
              }
            },
            "required": ["code", "message"]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "错误时间戳"
          },
          "requestId": {
            "type": "string",
            "description": "请求ID"
          }
        },
        "required": ["success", "error", "timestamp", "requestId"]
      }
    }
  }
}